src/main/webapp/app/entities/processo/update/processo-update.component.ts
import { Component, OnInit } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize, map } from 'rxjs/operators';

import { ProcessoFormService, ProcessoFormGroup } from './processo-form.service';
import { IProcesso } from '../processo.model';
import { ProcessoService } from '../service/processo.service';
import { ITurma } from 'app/entities/turma/turma.model';
import { TurmaService } from 'app/entities/turma/service/turma.service';
import { IEntregaDiploma } from 'app/entities/entrega-diploma/entrega-diploma.model';
import { EntregaDiplomaService } from 'app/entities/entrega-diploma/service/entrega-diploma.service';
import { StatusProcesso } from 'app/entities/enumerations/status-processo.model';
import { Status } from 'app/entities/enumerations/status.model';
import { Enviado } from 'app/entities/enumerations/enviado.model';

@Component({
  selector: 'jhi-processo-update',
  templateUrl: './processo-update.component.html',
})
export class ProcessoUpdateComponent implements OnInit {
  isSaving = false;
  processo: IProcesso | null = null;
  statusProcessoValues = Object.keys(StatusProcesso);
  statusValues = Object.keys(Status);
  enviadoValues = Object.keys(Enviado);

  turmasSharedCollection: ITurma[] = [];
  entregaDiplomasSharedCollection: IEntregaDiploma[] = [];

  editForm: ProcessoFormGroup = this.processoFormService.createProcessoFormGroup();

  constructor(
    protected processoService: ProcessoService,
    protected processoFormService: ProcessoFormService,
    protected turmaService: TurmaService,
    protected entregaDiplomaService: EntregaDiplomaService,
    protected activatedRoute: ActivatedRoute
  ) {}

  compareTurma = (o1: ITurma | null, o2: ITurma | null): boolean => this.turmaService.compareTurma(o1, o2);

  compareEntregaDiploma = (o1: IEntregaDiploma | null, o2: IEntregaDiploma | null): boolean =>
    this.entregaDiplomaService.compareEntregaDiploma(o1, o2);

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ processo }) => {
      this.processo = processo;
      if (processo) {
        this.updateForm(processo);
      }

      this.loadRelationshipsOptions();
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const processo = this.processoFormService.getProcesso(this.editForm);
    if (processo.id !== null) {
      this.subscribeToSaveResponse(this.processoService.update(processo));
    } else {
      this.subscribeToSaveResponse(this.processoService.create(processo));
    }
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<IProcesso>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(processo: IProcesso): void {
    this.processo = processo;
    this.processoFormService.resetForm(this.editForm, processo);

    this.turmasSharedCollection = this.turmaService.addTurmaToCollectionIfMissing<ITurma>(this.turmasSharedCollection, processo.turma);
    this.entregaDiplomasSharedCollection = this.entregaDiplomaService.addEntregaDiplomaToCollectionIfMissing<IEntregaDiploma>(
      this.entregaDiplomasSharedCollection,
      processo.entregaDiploma
    );
  }

  protected loadRelationshipsOptions(): void {
    this.turmaService
      .query()
      .pipe(map((res: HttpResponse<ITurma[]>) => res.body ?? []))
      .pipe(map((turmas: ITurma[]) => this.turmaService.addTurmaToCollectionIfMissing<ITurma>(turmas, this.processo?.turma)))
      .subscribe((turmas: ITurma[]) => (this.turmasSharedCollection = turmas));

    this.entregaDiplomaService
      .query()
      .pipe(map((res: HttpResponse<IEntregaDiploma[]>) => res.body ?? []))
      .pipe(
        map((entregaDiplomas: IEntregaDiploma[]) =>
          this.entregaDiplomaService.addEntregaDiplomaToCollectionIfMissing<IEntregaDiploma>(entregaDiplomas, this.processo?.entregaDiploma)
        )
      )
      .subscribe((entregaDiplomas: IEntregaDiploma[]) => (this.entregaDiplomasSharedCollection = entregaDiplomas));
  }
}
